+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/adapter/quartz/trigger/IndexScheduleTriggerConfig.java
@@ -43,11 +43,4 @@ public class IndexScheduleTriggerConfig {
 			.build();
 
 	}
-
-	@Bean
-	public SchedulerFactoryBeanCustomizer schedulerFactoryBean() {
-		return (SchedulerFactoryBeanCustomizer) schedulerFactoryBean -> {
-			schedulerFactoryBean.setGlobalJobListeners(new GlobalJobListener());
-		};
-	}
 }
